+++ psychopy/platform_specific/__init__.py
@@ -5,13 +5,26 @@
 
 import sys, platform
 
-def rush(value=False): 
-    #dummy method. Should be overridden by imports below if they exist and work
-    #if we return something it means rush doesn't work on this platform
+#dummy methods Should be overridden by imports below if they exist
+def rush(value=False):
+    """
+    """
+    #dummy method.
     return False
 def waitForVBL():
-    pass
-    
+    """DEPRECATED: waiting for a VBL is handled by the screen flip
+    """
+    return False
+def sendStayAwake():
+    """Sends a signal to your system to indicate that the computer is in use and
+    should not sleep. This should be sent periodically, but PsychoPy will send
+    the signal by default on each screen refresh.
+    Added: v1.79.00
+
+    Currently supported on: windows, OS X
+    """
+    return False
+
 if sys.platform=='win32':#NB includes vista and 7 (but not sure about vista64)
     from win32 import *
 elif sys.platform=='darwin':